requirements.txt
asn1crypto==0.24.0
attrs==18.2.0
Automat==0.7.0
backports.csv==1.0.7
cairocffi==1.0.2
certifi==2019.3.9
cffi==1.12.2
chardet==3.0.4
constantly==15.1.0
cryptography==2.4.2
cssselect==1.0.3
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.1.4
django-annoying==0.10.4
django-filter==0.15.3
django-import-export==1.2.0
djangorestframework==3.6.2
et-xmlfile==1.0.1
hyperlink==18.0.0
idna==2.8
incremental==17.5.0
jdcal==1.4.1
lxml==4.2.5
numpy==1.15.4
odfpy==1.4.0
openpyxl==2.6.2
pandas==0.23.4
parsel==1.5.1
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycparser==2.19
PyDispatcher==2.0.5
PyHamcrest==1.9.0
pyOpenSSL==18.0.0
python-dateutil==2.7.5
pytz==2018.9
PyYAML==5.1
queuelib==1.5.0
requests==2.21.0
Scrapy==1.5.1
service-identity==18.1.0
six==1.12.0
tablib==0.13.0
Twisted==18.9.0
urllib3==1.24.1
w3lib==1.19.0
xlrd==1.2.0
xlwt==1.3.0
zope.interface==4.6.0
